GH-105 - moved some tests to controller unit tests, as they do not need to spin up a mongodb
